[IMP] hr_recruitment_survey: expire survey links
There is no need to be able to access survey links when the applicant is
hired, deleted, archived or refused.
To do so, to mark surveys as expired, we decided to close their
deadlines when needed.

I made so that a new invite would be always regenerated / sent when
sending it via the wizard, to avoid confusion with the same invite being
resent sometimes.

Survey links would always refuse logged in users that were not the
intended recipient, even if the answer token was correct. Although
opening the url in a private window would allow us to access the survey.
This commit fixes that, allowing anyone to access the survey IF the
answer_token is valid in the URL.

task-4784231