[IMP] orm: make signaling part of the transaction
Move the implementation of cache and registry invalidation to the
Transaction object.

Creating a Transaction checks signaling (triggered from the creation of
the first Environment). Changes are signaled on commit of the cursor.
Transaction's `reset` method is called after commit or rollback, in case
we continue using the cursor (it is not "closing"), then we re-check
signaling.
This means that the following code handles all signaling correctly:

```py
with cursor() as cr:  # signal or reset on exit
  env = api.Environment(cr, ...)  # check signaling
  cr.commit()  # signal changes and reset transaction
  cr.rollback()  # reset changes and transaction
```