[FIX] html_builder, *: use hidden_element_predicates to skip siblings
*: website

The function `getNearestSibling` relied on whether an element had style
with `display: none` to tell whether it should be skipped when looking
for the next sibling.

For popup which are hidden when their child is invisible, this was
not working always well. It did work most of the time thanks to the
interaction that attempts to sync the `d-none` class on the popup with
the visibility of its modal.

With this commit, the function `getNearestSibling` skips elements which
match `is_element_hidden_predicates`

Steps to reproduce:
- Open website builder
- Add a section
- Add a popup
- Close the popup
- Click on the section
- Change an option on the section
- Show the popup by clicking the eye in "Invisible Elements" panel
- Undo
- Bug: there is the "Move down" arrow of the overlay buttons

Last commit improving the sync of `d-none`: 71dacea296e76d1ba18a8a5d44b20365eed9951c

task-5149984